[twlox] fix: Environment replaces a value instead of modifying it
Previously, the "Environment" struct replaced a value of the smart
pointer "Rc<RefCell<_>>" instead of modifying it. This led to errors
with global variables and assigning in general. When a value was
assigned to a variable, it completely overwrote the previous value.
Consequently, if there were more references to that variable, they
weren't updated.

Write a mock object for the resolver callback so we can test the full
functionality of the resolver without the hassle. This allows us to test
resolver functionality easier by predicting how many times and with what
data the resolver should call the callback, and then comparing it with
the actual data. This helped to find the bug.

Replace many ".into()" calls with a more explicit function, such as
".to_rc()" for "Value" or ".to_box()" for "Stmt" and "Expr". This
improves code readability.

Also, while testing the resolver, I noticed that we desugar grammar
during parsing, and as a result, I can't properly test the resolver with
a for loop because it's hard to determine the variable positions after
the desugaring.