[twlox] refactor: Resolver errors, use "Rc::clone()" convention
* Add "keyword" field to the "Return" statement to produce better error
message in the Resolver.

* Rework Resolver to include multiple errors and not exit the resolver
method until all child nodes are resolved. This way we can achieve a
better UX because the user will know more errors from single run.

* Use "thiserror" crate to reduce Error trait implementation boilerplate.

* Add Resolver test "this_keyword_in_method".

* Rework "compile_time_keywords_errors" test to ensure that Resolver
finds all errors in a single run. This test already helped detect a
bug where the Resolver exits the function prematurely (via "?") and the
".end_scope()" method is ignored as a result.

* Apply best practice to use "Rc::clone()" instead of ".clone()".
> By using "Rc::clone" for reference counting, we can visually
> distinguish between the deep-copy kinds of clones and the kinds of
> clones that increase the reference count.
Ref: https://doc.rust-lang.org/book/ch15-04-rc.html#using-rct-to-share-data