[twlox] refactor: define "print" as global function; improve cli usage
Before this "print" was a separate statement node in ast. Instead of
being a normal function, it was called like "print a". This makes no
sense when the interpreter already supports functions. Define "print" as
a global function and call it as a regular function "print(a)".

Improve cli UX. Add cli help message. Parse cli arguments with lexopt
(god level library).