[twlox] test: move ui tests from unit to integration tests
Before this, integration tests were written in the "runner" module, but
it looks weird. We need to move them to the integration tests and create
separate file with ".lox" extension for each test program. These types
of tests are called "ui" (I've seen them called that in rustc tests).
To be able to use "cargo test" to run all the tests and not reinvent the
wheel with a new testing framework we need to generate a test case for
each file. Only stdout is currently supported. I'm still in doubt about
the best way to support stderr, since it can contain many errors, and
usually only the first one is meaningful.